+++ src/cmd/ksh93/sh/nvdisc.c
@@ -894,8 +894,8 @@ int nv_clone(Namval_t *np, Namval_t *mp, int flags)
 {
 	Namfun_t	*fp, *fpnext;
 	const char	*val = mp->nvalue;
-	unsigned short	flag = mp->nvflag;
-	unsigned short	size = mp->nvsize;
+	uint32_t	flag = mp->nvflag;
+	uint32_t	size = mp->nvsize;
 	for(fp=mp->nvfun; fp; fp=fpnext)
 	{
 		fpnext = fp->next;
